third_party/WebKit/LayoutTests/fast/multicol/nested-balanced-inner-column-count-1-with-forced-break-2.html

<!DOCTYPE html>
<p>There should be a green square below, and no red.</p>
<div id="outer" style="columns:2; column-gap:0; width:20px; line-height:20px; background:red;">
    <div style="columns:1; column-gap:0; background:green;">
        <div><br></div>
        <div style="break-before:column;"><br></div>
    </div>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
    var elm = document.getElementById("outer");
    assert_equals(elm.offsetHeight, 20);
}, "Inner column count is 1");
</script>
